requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: only needed if you set GEMINI_API_KEY
google-genai>=1.0.0
